remove file/config split, we use just config.rs now
